xui: simplify frame sizing
Frame size options take float64 directly. Omitting an option uses the
default sizing behavior, so the Auto type is no longer needed.

Dimensions use float64 throughout. Frames track which options were
supplied separately so explicit zero dimensions remain meaningful.
Rendering plans reserve zero ideal dimensions for omitted declarations.

+++ xui/internal/fixture/fixture.go
@@ -261,7 +261,7 @@ func iconDemo() View {
 				Font(f)
 		}),
 		HStack(
-			For([]int{20, 30, 40}, nil, func(n int) View {
+			For([]float64{20, 30, 40}, nil, func(n float64) View {
 				return HStack(
 					Icon("film"),
 					Yellow.
